rpm: upgrade to version 4.14.2
Corresponding changes in few dependent specs.
Fixed a regression in valgrind.
Removed python dependency from rpm

Add size & installed package checks to docker image creation script.

Also took the changes from below commit for rpm macros.
https://github.com/vmware/photon/commit/6f03b51fcae78a3d9cfb2911f7e163dc90f8ce92

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/13601
Tested-by: gerrit-photon <photon-checkins@vmware.com>

+++ support/dockerfiles/photon/make-docker-image.sh
@@ -1,18 +1,13 @@
 #!/bin/bash
 
-if [ "$#" -lt 1 ]; then
-	echo "Script to create new photon base docker image."
-	echo "Usage: $0 <path to workspace>"
-	exit -1
-fi
+set -ex
 
-set -e
-set -x
+echoerr()
+{
+  echo -e "$*" 1>&2
+}
 
-PROGRAM=$0
-MAIN_PACKAGE=$1
-
-
+arch="$(uname -m)"
 TEMP_CHROOT=$(pwd)/temp_chroot
 ROOTFS_TAR_FILENAME=photon-rootfs-$PHOTON_RELEASE_VERSION-$PHOTON_BUILD_NUMBER.tar.bz2
 STAGE_DIR=$(pwd)/stage
@@ -22,7 +17,7 @@ rm -rf /etc/yum.repos.d/*
 cat > /etc/yum.repos.d/photon-local.repo <<- EOF
 
 [photon-local]
-name=VMware Photon Linux 1.0(x86_64)
+name=VMware Photon Linux 2.0($arch)
 baseurl=file://$(pwd)/stage/RPMS
 gpgkey=file:///etc/pki/rpm-gpg/RPM-GPG-KEY
 gpgcheck=0
@@ -31,16 +26,49 @@ skip_if_unavailable=True
 
 EOF
 
-rm -rf $TEMP_CHROOT 
+rm -rf $TEMP_CHROOT
 mkdir $TEMP_CHROOT
+
+tdnf install -y tar grep coreutils
 
 rpm --root $TEMP_CHROOT/ --initdb
 
-tdnf --installroot $TEMP_CHROOT/ --rpmverbosity 10 install -y filesystem bash toybox tdnf photon-release photon-repos
+tdnf --installroot $TEMP_CHROOT/ \
+     install -y \
+     filesystem bash toybox tdnf photon-release photon-repos
+
+actual_pkg_list=($(tdnf --installroot $TEMP_CHROOT/ \
+                        --disablerepo=* -q \
+                        list installed 2>/dev/null | cut -d' ' -f1))
+
+expected_pkg_list=(
+  bash.$arch bzip2-libs.$arch ca-certificates.$arch ca-certificates-pki.$arch
+  curl.$arch curl-libs.$arch e2fsprogs-libs.$arch elfutils-libelf.$arch
+  expat-libs.$arch filesystem.$arch glibc.$arch hawkey.$arch krb5.$arch
+  libcap.$arch libdb.$arch libgcc.$arch libsolv.$arch libssh2.$arch lua.$arch
+  ncurses-libs.$arch nspr.$arch nss-libs.$arch openssl.$arch
+  photon-release.noarch photon-repos.noarch popt.$arch readline.$arch
+  rpm-libs.$arch sqlite-libs.$arch tdnf.$arch tdnf-cli-libs.$arch
+  toybox.$arch xz-libs.$arch zlib.$arch
+  )
+
+actual_pkg_count=${#actual_pkg_list[@]}
+expected_pkg_count=${#expected_pkg_list[@]}
+
+pkg_diff="$(echo ${expected_pkg_list[@]} ${actual_pkg_list[@]} | \
+            tr ' ' '\n' | sort | uniq -u)"
+
+if [ $expected_pkg_count -ne $actual_pkg_count ] || [ -n "${pkg_diff}" ]; then
+  echoerr "Following package difference found in docker image:\n${pkg_diff}"
+  echoerr "Expected package count: $expected_pkg_count"
+  echoerr "Actual package count: $actual_pkg_count"
+  rm -rf $TEMP_CHROOT
+  exit 1
+fi
 
 rpm --root $TEMP_CHROOT/ --import $TEMP_CHROOT/etc/pki/rpm-gpg/*
 
-cd $TEMP_CHROOT
+pushd $TEMP_CHROOT
 # cleanup anything not needed inside rootfs
 rm -rf usr/src/
 rm -rf home/*
@@ -50,12 +78,20 @@ rm -rf var/log/*
 echo "export TERM=linux" >> etc/bash.bashrc
 
 #find var/cache/tdnf/photon/rpms -type f -name "*.rpm" -exec rm {} \;
-tdnf install -y tar
 tar cpjf ../$ROOTFS_TAR_FILENAME .
-mkdir -p $STAGE_DIR
-mv ../$ROOTFS_TAR_FILENAME $STAGE_DIR/
-cd ..
+popd
+
+max_size=$(( 14 * 1024 * 1024 ))
+actual_size=$(wc -c $ROOTFS_TAR_FILENAME | cut -d' ' -f1)
+if (( $actual_size > $max_size )); then
+  echoerr "ERROR: docker image tarball size is bigger than expected"
+  echoerr "Expected size(in bytes): $max_size"
+  echoerr "Actual size(in bytes): $actual_size"
+  rm -rf $TEMP_CHROOT $ROOTFS_TAR_FILENAME
+  exit 1
+fi
+
+mv $ROOTFS_TAR_FILENAME $STAGE_DIR/
 
 # cleanup
 rm -rf $TEMP_CHROOT
-
